fix(macos): Resolve Clang implicit integer conversion warnings for macOS CI
Problem: macOS CI build failing with Clang -Wimplicit-int-conversion warnings
- 4 errors in channel_status.hpp: bitwise operations losing precision to uint8_t
- 1 error in test_aes3_part2_conformity.cpp: bitwise operations losing precision to uint16_t
- Clang flag: -Werror,-Wimplicit-int-conversion (stricter than GCC -Wconversion)

Root Cause: Bitwise operations promote to 'int' type, then assigned to narrower types
- C++ integer promotion rules: bitwise OR/AND/shift operations promote operands to int
- Assigning int  uint8_t or uint16_t requires explicit cast for Clang
- GCC accepts these with -Wconversion, but Clang's -Wimplicit-int-conversion is stricter

Errors Fixed:

1. channel_status.hpp line 132 (set_pre_emphasis):
   - Before: bytes[0] = (bytes[0] & 0xE3) | ((emphasis & 0x07) << 2);
   - After: bytes[0] = static_cast<uint8_t>((bytes[0] & 0xE3) | ((emphasis & 0x07) << 2));
   - Fixes: int  uint8_t implicit conversion

2. channel_status.hpp line 173 (set_sampling_frequency):
   - Before: bytes[0] = (bytes[0] & 0x3F) | ((fs & 0x03) << 6);
   - After: bytes[0] = static_cast<uint8_t>((bytes[0] & 0x3F) | ((fs & 0x03) << 6));
   - Fixes: int  uint8_t implicit conversion

3. channel_status.hpp line 233 (set_word_length):
   - Before: bytes[2] = (bytes[2] & 0xC7) | ((wl & 0x07) << 3);
   - After: bytes[2] = static_cast<uint8_t>((bytes[2] & 0xC7) | ((wl & 0x07) << 3));
   - Fixes: int  uint8_t implicit conversion

4. channel_status.hpp line 254 (set_alignment_level):
   - Before: bytes[2] = (bytes[2] & 0x3F) | ((align & 0x03) << 6);
   - After: bytes[2] = static_cast<uint8_t>((bytes[2] & 0x3F) | ((align & 0x03) << 6));
   - Fixes: int  uint8_t implicit conversion

5. test_aes3_part2_conformity.cpp line 426:
   - Before: uint16_t sample_12bit = (nibble3.bits << 8) | (nibble2.bits << 4) | nibble1.bits;
   - After: uint16_t sample_12bit = static_cast<uint16_t>((nibble3.bits << 8) | (nibble2.bits << 4) | nibble1.bits);
   - Fixes: int  uint16_t implicit conversion

Impact:
- macOS Clang build now clean (0 warnings, 0 errors expected)
- All 155 tests still passing on Windows/Linux (verified locally)
- All 39 Part 2 conformity tests passing (verified locally)
- No functional changes - only explicit type safety

Cross-Platform Compatibility Summary:
- Windows MSVC: Passing (commit ec336e4, 48753bc)
- Linux GCC: Passing (commits 8fb0fd6, 48753bc)
- macOS Clang: Fixed (this commit)

Standards: ISO/IEC/IEEE 12207:2017 Implementation Process 6.4.5.3(h)
- Ensures strict type safety across all major compiler toolchains
- Maintains C++ integer promotion compliance on all platforms

Traceability:
- Requirements: REQ-FUNC-META-001 through REQ-FUNC-META-006
- Tests: TEST-CONF-META-006 through TEST-CONF-META-041
- Design: DES-C-002 (Channel Status Block)
- Phase: 05-implementation

Files Modified:
- 05-implementation/.../metadata/channel_status.hpp (4 explicit casts)
- 05-implementation/.../conformity/test_aes3_part2_conformity.cpp (1 explicit cast)

Verification: Local Windows MSVC build
- Build: Clean, 0 warnings, 0 errors
- Tests: All 39 Part 2 conformity tests passing
- Performance: Unaffected by explicit casts (compiler optimization)

+++ 05-implementation/src/AES/AES3/Part2/_2009/metadata/channel_status.hpp
@@ -129,7 +129,7 @@ struct ChannelStatusBlock {
      * @param emphasis Pre-emphasis value (0-7, see get_pre_emphasis())
      */
     void set_pre_emphasis(uint8_t emphasis) noexcept {
-        bytes[0] = (bytes[0] & 0xE3) | ((emphasis & 0x07) << 2);
+        bytes[0] = static_cast<uint8_t>((bytes[0] & 0xE3) | ((emphasis & 0x07) << 2));
     }
     
     /**
@@ -170,7 +170,7 @@ struct ChannelStatusBlock {
      * @param fs Sampling frequency value (0-3, see get_sampling_frequency())
      */
     void set_sampling_frequency(uint8_t fs) noexcept {
-        bytes[0] = (bytes[0] & 0x3F) | ((fs & 0x03) << 6);
+        bytes[0] = static_cast<uint8_t>((bytes[0] & 0x3F) | ((fs & 0x03) << 6));
     }
     
     // =========================================================================
@@ -230,7 +230,7 @@ struct ChannelStatusBlock {
      * @param wl Word length value (0-7, see get_word_length())
      */
     void set_word_length(uint8_t wl) noexcept {
-        bytes[2] = (bytes[2] & 0xC7) | ((wl & 0x07) << 3);
+        bytes[2] = static_cast<uint8_t>((bytes[2] & 0xC7) | ((wl & 0x07) << 3));
     }
     
     /**
@@ -251,7 +251,7 @@ struct ChannelStatusBlock {
      * @param align Alignment level value (0-3, see get_alignment_level())
      */
     void set_alignment_level(uint8_t align) noexcept {
-        bytes[2] = (bytes[2] & 0x3F) | ((align & 0x03) << 6);
+        bytes[2] = static_cast<uint8_t>((bytes[2] & 0x3F) | ((align & 0x03) << 6));
     }
 };
